Refactored repo to move /pkg to /internal/pkg
